+++ packages/kysely/src/keys/index.ts
@@ -2,13 +2,13 @@ import { Kysely } from "kysely";
 import { KeysAdapter } from "@authhero/adapter-interfaces";
 import { list } from "./list";
 import { create } from "./create";
-import { revoke } from "./revoke";
+import { update } from "./update";
 import { Database } from "../db";
 
 export function createKeysAdapter(db: Kysely<Database>): KeysAdapter {
   return {
     create: create(db),
     list: list(db),
-    revoke: revoke(db),
+    update: update(db),
   };
 }