+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/demo
 
+## 0.0.19
+
+### Patch Changes
+
+- authhero@0.2.21
+
 ## 0.0.18
 
 ### Patch Changes